Credit card information functionalities completed

+++ src/main/java/com/passwordManager/api/services/IdentityInfoServicesImpl.java
@@ -5,16 +5,16 @@ import com.passwordManager.api.data.models.Identity;
 import com.passwordManager.api.data.repositories.IdentityRepository;
 import com.passwordManager.api.dtos.requests.DeleteIdentityInfoRequest;
 import com.passwordManager.api.dtos.requests.EditIdentityInfoRequest;
+import com.passwordManager.api.dtos.requests.GetIdentityInfoRequest;
 import com.passwordManager.api.dtos.requests.IdentityInfoRequest;
-import com.passwordManager.api.exceptions.IdentityInfoNotFoundException;
-import com.passwordManager.api.exceptions.IncorrectNINnumberException;
-import com.passwordManager.api.exceptions.InvalidNiNException;
+import com.passwordManager.api.exceptions.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
 import static com.passwordManager.api.utilities.Mapper.map;
+import static com.passwordManager.api.utilities.NumericCipher.encrypt;
 
 @Service
 public class IdentityInfoServicesImpl implements IdentityInfoServices{
@@ -23,32 +23,42 @@ public class IdentityInfoServicesImpl implements IdentityInfoServices{
     @Override
     public Identity addIdentityInfo(IdentityInfoRequest identityInfoRequest) {
         validate(identityInfoRequest.getNationalIdentityNumber());
-        if (identityInfoRequest.getNationalIdentityNumber().length() != 11)throw new InvalidNiNException("Must contain 11 digits");
+        if (identityInfoRequest.getNationalIdentityNumber().length() != 11)throw new InvalidNiNException("NIN number must contain 11 digits");
         Identity identity = map(identityInfoRequest);
-
+        if (identityInfoRequest.getNationalIdentityNumber() != null)identity.setNationalIdentityNumber(encrypt(Long.parseLong(identityInfoRequest.getNationalIdentityNumber())));
         return identityRepository.save(identity);
     }
 
     @Override
     public Identity deleteIdentityInfo(DeleteIdentityInfoRequest deleteIdentityInfoRequest) {
-
+        if (deleteIdentityInfoRequest.getIdentityInfoId() == null) throw new PasswordManagerException("Identity does not exist");
         return identityRepository.deleteIdentityById(deleteIdentityInfoRequest.getIdentityInfoId());
     }
 
     @Override
     public Identity editIdentityInfo(EditIdentityInfoRequest editIdentityInfoRequest) {
-        validate(editIdentityInfoRequest.getNationalIdentityNumber());
+        if (editIdentityInfoRequest.getNationalIdentityNumber() != null)validate(editIdentityInfoRequest.getNationalIdentityNumber());
+        if (editIdentityInfoRequest.getIdentityInfoId() == null)throw new FieldRequiredException(
+                "Identity information id not specified");
         Optional<Identity> identity = identityRepository.findById(editIdentityInfoRequest.getIdentityInfoId());
         if (identity.isEmpty())throw new IdentityInfoNotFoundException("Identity does not exist");
-        Identity foundIdentity = map(editIdentityInfoRequest, identity);
+        Identity foundIdentity = map(editIdentityInfoRequest, identity.get());
         return identityRepository.save(foundIdentity);
     }
 
+    @Override
+    public Identity getIdentityInfo(GetIdentityInfoRequest getIdentityInfoRequest) {
+        if (getIdentityInfoRequest.getIdentityInfoId() == null)throw new FieldRequiredException(
+                "Identity information id not specified");
+        Optional<Identity> identity = identityRepository.findById(getIdentityInfoRequest.getIdentityInfoId());
+        if (identity.isEmpty())throw new IdentityInfoNotFoundException("Identity does not exist");
+        return identity.get();
+    }
 
 
-    private void validate(String NinNumber){
-        for (int position = 0; position < NinNumber.length(); position++) {
-            char character = NinNumber.charAt(position);
+    private void validate(String ninNumber){
+        for (int position = 0; position < ninNumber.length(); position++) {
+            char character = ninNumber.charAt(position);
             if (Character.isLetter(character))throw new IncorrectNINnumberException("The NIN " +
                     "number you entered is not valid");
         }
